#13 refactor: content 패키지 정리
