+++ src/main/java/com/example/sharemind/nonRealtimeConsult/domain/NonRealtimeConsult.java
@@ -5,7 +5,7 @@ import com.example.sharemind.customer.domain.Customer;
 import com.example.sharemind.global.common.BaseEntity;
 import com.example.sharemind.global.content.ConsultCategory;
 import com.example.sharemind.nonRealtimeConsult.content.NonRealtimeConsultStatus;
-import com.example.sharemind.nonRealtimeMessage.domain.NonRealtimeMessageType;
+import com.example.sharemind.nonRealtimeMessage.content.NonRealtimeMessageType;
 import com.example.sharemind.nonRealtimeMessage.exception.NonRealtimeMessageErrorCode;
 import com.example.sharemind.nonRealtimeMessage.exception.NonRealtimeMessageException;
 import jakarta.persistence.*;
